challenges/challenge-17/cookmaster/heater/requirements.txt
msgpack==1.0.7
packaging==23.2
python-can==4.3.1
typing_extensions==4.9.0
wrapt==1.16.0
